+++ src/core/tree/default_tree_vector.rs
@@ -25,19 +25,19 @@ impl <T: Default + Copy> TreeVec<T> for DefaultTreeVec<T> {
         index as i32
     }
 
-    fn get(self: &mut DefaultTreeVec<T>, index: u64) -> TreeNode<T> {
+    fn get(self: &mut DefaultTreeVec<T>, index: i32) -> TreeNode<T> {
         self.data[index as usize]
     }
 
-    fn swap(&mut self, index1: u64, index2: u64) {
+    fn swap(&mut self, index1: i32, index2: i32) {
         let node1 = self.data[index1 as usize];
         let node2 = self.data[index2 as usize];
         self.data[index1 as usize] = node2;
         self.data[index2 as usize] = node1;
     }
 
-    fn remove(&mut self, index: u64) {
-        self.empty.push(index);
+    fn remove(&mut self, index: i32) {
+        self.empty.push(index as u64);
         self.data[index as usize] = TreeNode::default();
     }
